% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absmax.R
\name{absmax}
\alias{absmax}
\title{Maximum of Absolute Values of Vector}
\usage{
absmax(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector or array.}

\item{na.rm}{A logical indicating whether missing values should be removed.}
}
\value{
The maximum of the absolute values of elements of the input vector.
}
\description{
Take the maximum of the absolute values of an input vector.
}
\examples{
x <- c(5, 3, -9, -100, 3.14159, 7.5)
absmax(x)
}

