% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_names.R
\name{attrnames}
\alias{attrnames}
\title{Get Names of Attributes}
\usage{
attrnames(obj)
}
\arguments{
\item{obj}{Any object.}
}
\value{
Vector of character strings with the names of the attributes.
}
\description{
Get the names of the attributes of an input object.
}
\examples{
x <- matrix(1:100, ncol = 5)
colnames(x) <- LETTERS[1:5]
attrnames(x)
}
