% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R
\name{bioc}
\alias{bioc}
\title{Conveniently Load Bioconductor}
\usage{
bioc()
}
\description{
Load Bioconductor's \code{biocLite} function by wrapping \code{source}.
}
\details{
Wrapper to \code{source("https://bioconductor.org/biocLite.R")}
}
