% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cd}
\alias{cd}
\title{A Function to Change Directories using UNIX Shorthand}
\usage{
cd(dir)
}
\arguments{
\item{dir}{Where do you want to go relative to the home directory? 
No defaults, so the target path must be specified in full.
Tilde expansion is performed when possible (OS X and Linux).}
}
\description{
This function allows you to change directories with 'cd'.
}
\examples{
cd()
}

