% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear}
\alias{clear}
\title{Clear the Current Screen/Buffer}
\usage{
clear()
}
\description{
Clear the screen with a call to \code{\link[base]{system}} and \code{clear}.
}
\details{
This function is merely a call to \code{system("clear")}
}
\examples{
\donttest{
system("clear")
}

}
