% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commas.R
\name{commas}
\alias{commas}
\title{Add Commas to a Large Number}
\usage{
commas(numbers)
}
\arguments{
\item{numbers}{Vector of non-negative numbers (will be rounded to integers)}
}
\value{
Character string with numbers written like \code{"5,771,009"}.
}
\description{
Convert a number to a string, with commas inserted at every 3rd digit.
}
\examples{
commas(c(2300, 9000, 21456, 987654890, 1256787, 345765, 1432))
}
