% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plots.R
\name{diaglmPlot_gg}
\alias{diaglmPlot_gg}
\title{Linear Model Diagnostic Plots with ggplot2}
\usage{
diaglmPlot_gg(model)
}
\arguments{
\item{model}{A linear model object produced by \code{lm()} or \code{glm()}.}
}
\description{
Produce standard diagnostic plots for linear models using ggplot2.
}
\examples{
n <- 10; x1 <- rnorm(n); y1 <- rnorm(n)
linmod <- lm(y1 ~ x1)
diaglmPlot_gg(linmod)
}

