% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{exit}
\alias{exit}
\title{Exit R Without Saving}
\usage{
exit()
}
\description{
Exit R without saving workspace, using the ubiquitous UNIX syntax.
}
\details{
This function is merely a call to \code{q("no")}.
}

