% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_num.R
\name{fac_num}
\alias{fac_num}
\title{Convert a Factor to Numeric}
\usage{
fac_num(x)
}
\arguments{
\item{x}{A vector containing a factor with numeric levels.}
}
\value{
The input factor made into a numeric vector.
}
\description{
Convert a factor with numeric levels to a non-factor (numeric).
}
\examples{
x <- factor(c(3, 4, 9, 4, 9), levels=c(3,4,9))
fac_num(x)
}

