% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_to_numeric.R
\name{factor_to_num}
\alias{factor_to_num}
\title{Convert a Factor to Numeric}
\usage{
factor_to_num(x)
}
\arguments{
\item{x}{A vector containing a factor with numeric levels.}
}
\value{
The input factor made into a numeric vector.
}
\description{
Convert a factor with numeric levels to a non-factor (numeric).
}
\examples{
x <- factor(c(3, 4, 9, 4, 9), levels = c(3, 4, 9))
factor_to_num(x)
}
