% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/len_uniq.R
\name{len_uniq}
\alias{len_uniq}
\title{Number of unique values}
\usage{
len_uniq(vec, na.rm = TRUE)
}
\arguments{
\item{vec}{A vector of any type.}

\item{na.rm}{If \code{TRUE}, remove any missing values}
}
\value{
Number of unique values.
}
\description{
Get the number of unique values in a input vector.
}
\examples{
x <- c(1, 3, 1, 1, NA, 2, 2, 3, NA, NA, 1, 3, 1)
len_uniq(x)
len_uniq(x, na.rm=FALSE)
}
\keyword{}

