% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lm_plot}
\alias{lm_plot}
\title{Linear Model Diagnostic Plots}
\usage{
lm_plot(x, ...)
}
\arguments{
\item{x}{A linear model object produced by \code{lm()}.}

\item{...}{Extra arguments, currently ignored.}
}
\description{
Produce standard diagnostic plots for linear models using ggplot2.
}
\examples{
n <- 100
x1 <- rnorm(n)
y1 <- rnorm(n)
linmod <- lm(y1 ~ x1)
plot(linmod)

}
