% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lmPlots_gg}
\alias{lmPlots_gg}
\title{Linear Model Diagnostic Plots with ggplot2}
\usage{
lmPlots_gg(model)
}
\arguments{
\item{model}{A linear model object produced by \code{lm()}.}
}
\description{
Produce standard diagnostic plots for linear models using ggplot2.
}
\examples{
n <- 100; x1 <- rnorm(n); y1 <- rnorm(n);
linmod <- lm(y1 ~ x1)
lmPlots_gg(linmod)
}
