% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newpkgs.R
\name{newpkg}
\alias{newpkg}
\title{New Package Skeleton}
\usage{
newpkg(name = "new_pkg")
}
\arguments{
\item{name}{- Character for the name of the new package, to be passed
directly to \code{devtools::create()}.}
}
\description{
Generate a skeleton for new R packages by invoking a series of utilities from
the Devtools package. This is merely a convenience utility for creating a new
directory with minimal package contents.
}
