% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newprojs.R
\name{newproj}
\alias{newproj}
\title{New Project Skeleton}
\usage{
newproj(name = "ProjectTemplate", minimal = TRUE, ...)
}
\arguments{
\item{name}{(character) - the name to be given to the new project directory}

\item{minimal}{(character) - option to set up only a minimal project
directory, passed to \code{ProjectTemplate::create.project()}}

\item{...}{- options to be passed to \code{ProjectTemplate::create.project()}}
}
\description{
Generate a skeleton for a data analysis project by calling ProjectTemplate
}
