% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{openfile}
\alias{openfile}
\title{Open a File}
\usage{
openfile(file)
}
\arguments{
\item{file}{File name (as character string).}
}
\description{
Open a file using \code{\link[base]{system}} and \code{open}.
}
\details{
I'd thought that to open a file you'd use \code{open} in MacOS and
\code{start} in Windows, but \code{system("start myfile.pdf")} doesn't work
work in Windows, but rather \code{system("open myfile.pdf")} does.
}
\examples{
\dontrun{openfile("myplot.pdf")}
}

