% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plots.R
\name{qqlmPlot_gg}
\alias{qqlmPlot_gg}
\title{Quantile-Quantile Plots with ggplot2}
\usage{
qqlmPlot_gg(x, distribution = "norm", ..., line.estimate = NULL,
  conf = 0.95, labels = names(x))
}
\arguments{
\item{x}{A numeric vector of residuals from a generalized linear model.}

\item{distribution}{The reference probability distribution for residuals.}
}
\description{
Produce standard quantile-quantile plots for modeling using ggplot2.
}
\examples{
n <- 10; x1 <- rnorm(n); y1 <- rnorm(n)
linmod <- lm(y1 ~ x1)
x <- linmod$residuals
qqlmPlot_gg(x)
}

