% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nima.R
\name{scale_color_nima}
\alias{scale_color_nima}
\title{Nima's ggplot2 theme - supplement: scale_color}
\usage{
scale_color_nima(...)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2]{ggplot}}}
}
\description{
Nima's ggplot2 theme scale_color supplement: colors optimized via ColorBrewer
}
