% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nima.R
\name{scale_fill_nima}
\alias{scale_fill_nima}
\title{Nima's ggplot2 theme - supplement: scale_fill}
\usage{
scale_fill_nima(...)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2]{ggplot}}}
}
\description{
Nima's ggplot2 theme scale_fill supplement: colors optimized via ColorBrewer
}
