% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_tools.R
\name{summarize_sim}
\alias{summarize_sim}
\title{Summarize Simulations Results}
\usage{
summarize_sim(simulation_results, truth, ci_level = 0.95)
}
\arguments{
\item{simulation_results}{...}

\item{truth}{...}

\item{ci_level}{...}
}
\description{
Summarize Simulations Results
}
\examples{
n_sim <- 100
n_obs <- c(100, 10000)
mu <- 2
sim_results <- lapply(n_obs, function(sample_size) {
  estimator_sim <- lapply(seq_len(n_sim), function(iter) {
    y_obs <- rnorm(sample_size, mu)
    est_param <- mean(y_obs)
    est_var <- var(y_obs)
    estimate <- tibble::as_tibble(list(
      param_est = est_param,
      param_var = est_var
    ))
    return(estimate)
  })
  estimates <- do.call(rbind, estimator_sim)
  return(estimates)
})
sim_summary <- lapply(sim_results, summarize_sim, truth = mu)
#
}
