% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlots.R
\name{survPlot_gg}
\alias{survPlot_gg}
\title{Kaplan-Meier Plot with ggplot2}
\usage{
survPlot_gg(s, CI = "def", plot.cens = TRUE, surv.col = "gg.def",
  cens.col = "red", lty.est = 1, lty.ci = 2, cens.shape = 3,
  back.white = FALSE, xlab = "Time", ylab = "Survival", main = "")
}
\arguments{
\item{s}{Survival model object, generated by methods like Kaplan-Meier.}

\item{CI}{Type of confidence interval for the survival object.}

\item{plot.cens}{Should the plot include the censored values?}

\item{surv.col}{Color for observations with outcomes of "survived".}

\item{cens.col}{Color for observations with outcomes of "censored".}

\item{lty.est}{The number of survival curves to be generated from the model.}

\item{lty.ci}{The number of lines to be generated for confidence intervals.}

\item{cens.shape}{The shapes of points plotted, passed to \code{geom_point}}

\item{back.white}{Boolean for desired background color in plots.}

\item{xlab}{A label for the x-axis, defaults to "Time".}

\item{ylab}{A label for the y-axis, defaults to "Survival".}

\item{main}{A main label for the survival plot, no default.}
}
\description{
Produce a survival plot of a Kaplan-Meier estimator using ggplot2.
}
\examples{
time <- c(5,6,8,3,22)
age <- c(46,35,30,30,36)
drug <- c(0,1,1,1,0)
censor <- c(1,0,1,1,1)
survdat <- as.data.frame(cbind(time,age,drug,censor))
s <- survfit(Surv(time, censor) ~ 1, conf.type = "none")
survPlot_gg(s)
}

