% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jetblack.R
\name{theme_jetblack}
\alias{theme_jetblack}
\title{A jet black ggplot2 theme with inverted colors}
\usage{
theme_jetblack(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}

\item{...}{Passed to \code{\link[ggplot2]{theme}}}
}
\value{
An object as returned by \code{\link[ggplot2]{theme}}
}
\description{
A jet black ggplot2 theme with inverted colors
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(y = mpg, x = disp, color = factor(cyl)))
p <- p + geom_point() + theme_jetblack()
p

}
\seealso{
\code{\link[ggplot2]{theme}}
}
