% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nima.R
\name{theme_nima}
\alias{nima_theme}
\alias{theme_nima}
\title{Nima's ggplot2 theme}
\usage{
theme_nima(base_size = 12, base_family = "", ...)

nima_theme(base_size = 12, base_family = "", ...)
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}

\item{...}{Passed to \code{\link[ggplot2]{theme}}}
}
\value{
An object as returned by \code{\link[ggplot2]{theme}}
}
\description{
Nima's ggplot2 theme: black border and no ticks
}
\examples{
library(ggplot2)
mtcars$cyl <- factor(mtcars$cyl)
p <- ggplot(mtcars, aes(y=mpg, x = disp, color = cyl)) 
p <- p + geom_point() + theme_nima()
p

}
\seealso{
\code{\link[ggplot2]{theme}}
}

