% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqlen.R
\name{uniqlen}
\alias{uniqlen}
\title{Find Number of Unique Values}
\usage{
uniqlen(vec, na.rm = TRUE)
}
\arguments{
\item{vec}{A vector of any type.}

\item{na.rm}{If \code{TRUE}, remove missing values.}
}
\value{
Number of unique values.
}
\description{
Get the number of unique values in an input vector.
}
\examples{
x <- c(1, 3, 1, 1, NA, 2, 2, 3, NA, NA, 1, 3, 1)
uniqlen(x)
uniqlen(x, na.rm = FALSE)
}
