% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{openfile}
\alias{openfile}
\title{Open a File}
\usage{
openfile(file)
}
\arguments{
\item{file}{File name (as character string).}
}
\description{
Open a file using \code{\link[base]{system}} and \code{open}.
}
\details{
Open files from R by using the default operating system program.
}
\examples{
\dontrun{
openfile("myplot.pdf")
}

}
