% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plots.R
\name{survPlot_gg}
\alias{survPlot_gg}
\title{Kaplan-Meier Plot with ggplot2}
\usage{
survPlot_gg(s, CI = "def", plot.cens = TRUE, surv.col = "gg.def",
  cens.col = "red", lty.est = 1, lty.ci = 2, cens.shape = 3,
  back.white = FALSE, xlab = "Time", ylab = "Survival", main = "")
}
\arguments{
\item{s}{___}

\item{CI}{___}

\item{plot.cens}{___}

\item{surv.col}{___}

\item{cens.col}{___}

\item{lty.est}{___}

\item{lty.ci}{___}

\item{cens.shape}{___}

\item{back.white}{___}

\item{xlab}{___}

\item{ylab}{___}

\item{main}{___}
}
\description{
Produce a survival plot of a Kaplan-Meier estimator using ggplot2.
}
\examples{
s <- km
survPlot_gg(s)
}

