% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plots.R
\name{survPlot_gg}
\alias{survPlot_gg}
\title{Kaplan-Meier Plot with ggplot2}
\usage{
survPlot_gg(s, CI = "def", plot.cens = TRUE, surv.col = "gg.def",
  cens.col = "red", lty.est = 1, lty.ci = 2, cens.shape = 3,
  back.white = FALSE, xlab = "Time", ylab = "Survival", main = "")
}
\arguments{
\item{s}{Survival model object, generated by methods like Kaplan-Meier.}

\item{CI}{Type of confidence interval for the survival object.}

\item{surv.col}{Color for observations with outcomes of "survived".}

\item{cens.col}{Color for observations with outcomes of "censored".}

\item{back.white}{Boolean for desired background color in plots.}

\item{xlab}{A label for the x-axis, defaults to "Time".}

\item{ylab}{A label for the y-axis, defaults to "Survival".}

\item{main}{A main label for the survival plot, no default.}
}
\description{
Produce a survival plot of a Kaplan-Meier estimator using ggplot2.
}
\examples{
s <- km
survPlot_gg(s)
}

