% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nima.R
\name{theme_nima}
\alias{theme_nima}
\alias{nima_theme}
\title{Nima's ggplot2 theme}
\usage{
theme_nima(base_size = 14, base_family = "Helvetica")

nima_theme(base_size = 14, base_family = "Helvetica")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}

\item{...}{Passed to \code{\link[ggplot2]{theme}}}
}
\value{
An object as returned by \code{\link[ggplot2]{theme}}
}
\description{
Nima's ggplot2 theme: white background, colors optimized
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(y = mpg, x = disp, color = factor(cyl)))
p <- p + geom_point() + scale_fill_nima() + scale_color_nima() + theme_nima()
p

}
\seealso{
\code{\link[ggplot2]{theme}}
}
